st up routes
